fix: preserve float precision in bbox coordinates for inference
Remove premature int() casts when storing bbox coordinates in
predict_receipt_from_dynamo(). The raw float coordinates are now
passed to _normalize_box_from_extents(), which handles rounding
and scaling itself. This matches the Swift BboxNormalizer behavior
and avoids truncation of fractional precision.

+++ receipt_layoutlm/receipt_layoutlm/inference.py
@@ -392,7 +392,8 @@ class LayoutLMInference:
         words = details.words
 
         # Group words by line and compute extents
-        by_line: Dict[int, List[Tuple[int, str, List[int]]]] = {}
+        # Store raw float coordinates; normalization handles rounding
+        by_line: Dict[int, List[Tuple[int, str, List[float]]]] = {}
         max_x = 0.0
         max_y = 0.0
         for w in words:
@@ -400,7 +401,7 @@ class LayoutLMInference:
             max_x = max(max_x, x1)
             max_y = max(max_y, y1)
             by_line.setdefault(w.line_id, []).append(
-                (w.word_id, w.text, [int(x0), int(y0), int(x1), int(y1)])
+                (w.word_id, w.text, [x0, y0, x1, y1])
             )
 
         tokens_per_line: List[List[str]] = []
